drawNode cleaned up. Actors now keep their avatar on creation.
