--- frontend/public/js/graphics/draw_origin_table.js
@@ -1,7 +1,0 @@
-'use strict';
-
-var menuBuilder = require('../menu_builder');
-
-module.exports = function(container, menu, map) {
-    return menu;
-}
